chore: `SidebarCollapser` component
